+++ release-notes.md
@@ -13,6 +13,9 @@ For mod developers:
 ## 1.12
 See [log](https://github.com/Pathoschild/SMAPI/compare/1.11...1.12).
 
+For players:
+* The installer now lets you choose the install path if you have multiple copies of the game, instead of using the first path found.
+
 For mod developers:
 * Unknown mod manifest fields are now stored in `IManifest::ExtraFields`.
 * The content API now defaults to `ContentSource.ModFolder`.